Only clone a message when necessary
This fixes a bug where a msg is cloned even though it's a different
message from any others e.g., `node.send([msg, {...msg}])` will clone
the 2nd msg despite it holding a different reference to any other
messages.

By not cloning the message, it's possible to use references to the msg
objects to store data privately across the flow.